+++ src/Exceptions/ValidationException.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace TPG\PayFast\Exceptions;
+
+class ValidationException extends \Exception
+{
+
+}